Kind of fixed the progress bar bug
When an element is styled with visibility: hidden it still takes up
space on the page so it was triggering a gesture when the progress bar
was clicked on. The display property should be used instead. The
problem is since we had to nest the progress divs they now vanish with
bookmarks.